todomvc-django/requirements.txt
defusedxml==0.4.1
Django==1.10.4
django-allauth==0.29.0
django-rest-auth==0.8.2
djangorestframework==3.5.3
gunicorn==19.6.0
oauthlib==2.0.1
psycopg2==2.6.2
python3-openid==3.0.10
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
whitenoise==3.2.2
